+++ dist.js
@@ -19,16 +19,16 @@
   _exports.default = _default;
 
   function _default() {
-    // elements
+    // root elements
     const root = document.querySelector('[data-ccs="root"]');
     const themeMenu = document.querySelector('[data-ccs="menu"]');
     const invertBtn = document.querySelector('[data-ccs-input="mode"]');
     const modeLight = document.querySelector('[data-ccs-input="light-mode"]');
     const modeDark = document.querySelector('[data-ccs-input="dark-mode"]');
     const modeAuto = document.querySelector('[data-ccs-input="auto-mode"]');
-    const unsetBtn = document.querySelector('[data-ccs-input="unset"]'); // selecting elements
+    const unsetBtn = document.querySelector('[data-ccs-input="unset"]'); // controls
 
-    const selectElements = {
+    const controls = {
       theme: document.querySelector('[data-ccs-input~="theme"]'),
       hue: document.querySelector('[data-ccs-input="hue"]'),
       sat: document.querySelector('[data-ccs-input="saturation"]'),
@@ -38,10 +38,31 @@
 
     const showUnsetBtn = (show = true) => {
       if (unsetBtn) {
-        show ? unsetBtn.removeAttribute('hidden') : unsetBtn.setAttribute('hidden', '');
+        show ? unsetBtn.removeAttribute("hidden") : unsetBtn.setAttribute("hidden", "");
       }
-    }; // set a value
+    }; // attributes
 
+
+    const attrs = {
+      theme: "data-ccs-theme"
+    }; // properties
+
+    const props = {
+      hue: "--ccs-prime--user",
+      sat: "--ccs-s--user",
+      light: "--ccs-l--user",
+      contrast: "--ccs-contrast--user",
+      mode: "--ccs-mode--user"
+    }; // local storage
+
+    const store = {
+      theme: "ccsTheme",
+      mode: "ccsMode",
+      hue: "ccsHue",
+      sat: "ccsSat",
+      light: "ccsLight",
+      contrast: "ccsContrast"
+    }; // set a value
 
     const setValue = (type, to, toStore = true) => {
       if (to) {
@@ -56,48 +77,6 @@
           showUnsetBtn();
         }
       }
-    }; // attributes
-
-
-    const attrs = {
-      theme: 'data-ccs-theme'
-    }; // properties
-
-    const props = {
-      hue: '--ccs-prime--user',
-      sat: '--ccs-s--user',
-      light: '--ccs-l--user',
-      contrast: '--ccs-contrast--user',
-      mode: '--ccs-mode--user'
-    }; // local storage
-
-    const store = {
-      theme: 'ccsTheme',
-      mode: 'ccsMode',
-      hue: 'ccsHue',
-      sat: 'ccsSat',
-      light: 'ccsLight',
-      contrast: 'ccsContrast'
-    };
-    const subTheme = ['hue', 'sat', 'light', 'contrast'];
-
-    const resetSelect = (sub = null) => Object.keys(selectElements).forEach(type => {
-      if (!sub || sub.includes(type)) {
-        const el = selectElements[type];
-        selectElements[type].value = el.getAttribute('data-default');
-      }
-    });
-
-    const setSelection = (type, selection) => {
-      const unsetTheme = selectElements.theme?.dataset.ccsInput.includes('unset-values');
-
-      if (type === 'theme' && unsetTheme) {
-        clearProps(subTheme);
-        clearStore(subTheme);
-        resetSelect(subTheme);
-      }
-
-      setValue(type, selection);
     };
 
     const clearProps = (sub = null) => Object.keys(props).forEach(prop => {
@@ -110,14 +89,41 @@
       if (!sub || sub.includes(type)) {
         localStorage.removeItem(store[type]);
       }
-    }); // clear all settings on reset
+    });
+
+    const resetControls = (sub = null) => Object.keys(controls).forEach(type => {
+      if (!sub || sub.includes(type)) {
+        const el = controls[type];
+
+        if (el) {
+          el.value = el.getAttribute("data-default");
+        }
+      }
+    });
+
+    const subTheme = ["hue", "sat", "light", "contrast"];
+
+    const setSelection = (type, selection) => {
+      const unsetTheme = controls.theme && controls.theme.dataset.ccsInput && controls.theme.dataset.ccsInput.includes("unset-values");
+
+      if (type === "theme" && unsetTheme) {
+        clearProps(subTheme);
+        clearStore(subTheme);
+        resetControls(subTheme);
+      }
+
+      setValue(type, selection);
+    }; // clear all settings on reset
 
 
     const unset = () => {
-      setValue('theme', selectElements.theme.getAttribute('data-default'), false);
+      if (controls.theme) {
+        setValue("theme", controls.theme.getAttribute("data-default"), false);
+      }
+
       clearStore();
       clearProps();
-      resetSelect();
+      resetControls();
       showUnsetBtn(false);
 
       if (modeAuto) {
@@ -126,9 +132,7 @@
     }; // modes
 
 
-    const getMode = () => {
-      return Number(getComputedStyle(root).getPropertyValue('--ccs-mode').trim());
-    };
+    const getMode = () => parseInt(getComputedStyle(root).getPropertyValue("--ccs-mode").trim(), 10);
 
     const changeMode = scheme => {
       const schemeDict = {
@@ -139,11 +143,11 @@
       const setting = schemeDict[scheme];
 
       if (setting) {
-        setValue('mode', setting);
+        setValue("mode", setting);
       } else {
         // if auto, remove mode props from root and store
-        localStorage.removeItem('ccsMode');
-        root.style.removeProperty(props.mode);
+        clearStore(["mode"]);
+        clearProps(["mode"]);
       }
     };
 
@@ -152,26 +156,19 @@
         1: modeLight,
         [-1]: modeDark
       };
-      setValue('mode', getMode() * -1);
+      const mode = getMode();
+      setValue("mode", mode * -1);
+      const modeBtn = modeDict[mode * -1];
 
-      if (modeLight || modeDark) {
-        modeDict[getMode()].checked = true;
+      if (modeBtn) {
+        modeBtn.checked = true;
       }
     }; // initialize everything
 
 
     const initMenu = () => {
-      themeMenu.removeAttribute('hidden');
-    };
-
-    const initValue = type => {
-      selectElements[type].setAttribute('data-default', selectElements[type].value);
-      const to = localStorage.getItem(store[type]);
-
-      if (to) {
-        setValue(type, to, false);
-        selectElements[type].value = to;
-        showUnsetBtn();
+      if (themeMenu) {
+        themeMenu.removeAttribute("hidden");
       }
     };
 
@@ -183,29 +180,43 @@
           1: modeLight,
           [-1]: modeDark
         };
-        setValue('mode', to);
-        showUnsetBtn();
-        modeDict[to].checked = true;
+        const modeBtn = modeDict[to];
+        setValue("mode", to);
+
+        if (modeBtn) {
+          modeBtn.checked = true;
+        }
       } else if (modeAuto) {
         modeAuto.checked = true;
+      }
+    };
+
+    const initValue = type => {
+      controls[type].setAttribute("data-default", controls[type].value);
+      const to = localStorage.getItem(store[type]);
+
+      if (to) {
+        setValue(type, to, false);
+        controls[type].value = to;
+        showUnsetBtn();
       }
     };
     /* init defaults */
 
 
-    document.onload = initMenu();
-    document.onload = initMode();
-    /* attach eventlistners */
+    window.onload = initMenu();
+    window.onload = initMode();
+    /* attach event listeners */
 
-    invertBtn && invertBtn.addEventListener('click', () => toggleMode());
-    modeLight && modeLight.addEventListener('change', () => changeMode('light'));
-    modeDark && modeDark.addEventListener('change', () => changeMode('dark'));
-    modeAuto && modeAuto.addEventListener('change', () => changeMode('auto'));
-    unsetBtn && unsetBtn.addEventListener('click', () => unset());
-    Object.keys(selectElements).forEach(type => {
-      if (selectElements[type]) {
-        document.onload = initValue(type);
-        selectElements[type].addEventListener('change', e => setSelection(type, e.target.value));
+    invertBtn && invertBtn.addEventListener("click", toggleMode);
+    modeLight && modeLight.addEventListener("change", () => changeMode("light"));
+    modeDark && modeDark.addEventListener("change", () => changeMode("dark"));
+    modeAuto && modeAuto.addEventListener("change", () => changeMode("auto"));
+    unsetBtn && unsetBtn.addEventListener("click", unset);
+    Object.keys(controls).forEach(type => {
+      if (controls[type]) {
+        window.onload = initValue(type);
+        controls[type].addEventListener("change", e => setSelection(type, e.target.value));
       }
     });
   }